Create MessageReceiver and MessageDispatcher utils (#175)
* Create MessageReceiver and MessageDispatcher utils

We now have three bus implementations that each create a web server to
receiver messages, do something with that message, and then dispatch the
message to subscribers. The MessageReceiver and MessageDispatcher utils
allow a bus to focus on the value add logic of the bus and share more of
the plumbing.

Note: I'm using the term message here instead of event since we don't
require that all traffic on a bus be CloudEvents. The Message struct is
capable of transporting a CloudEvent without limiting other cases.

The stub bus has been updated to use these new helpers. The gcppubsub
and kafka buses still need to be updated.

* Support header prefix whitelisting

* Update GCP Pub/Sub bus to use message receiver/dispatcher

* Update Kafka bus to use message receiver/dispatcher

* Review feedback