Handle github webhook missing (#252)
* Use log.Printf instead of glog

glog doesn't emit to stdout or stderr without flags. This makes it
unsuitable for use by a wrapped function with a standardized execution
environment. log.Printf omits levels but emits to stdout, making it
better for this use case.

* Consider missing webhook error a success

If the webhook doesn't exist, the final state is what we want even if we
don't know how it got there.