Subscription controller tweaks (#459)
* Correct some comments and variable names

* Add IsReady and Mark* funcs to SubscriptionStatus

These make the controller slightly easier to read.

Assuming that the Ready condition flip to true is intended to be
a FromReady flip.

* Remove GetSpecJSON

Replaced with Generation duck type.

* Add IgnoreTimes option to controller table tests

Time fields in conditions often need to be ignored because they rarely
compare as equal. Setting IgnoreTimes true in the test case will cause
diffs to ignore apis.VolatileTime values.

* Update test to expect initialized condition

All conditions are now defaulted to Unknown on first reconcile.

* Correct vet issue in table test helper