Fix context threading for Broker signal handling. (#2792)
The Broker filter libraries were discarding the signal-aware context they were passed in, which meant that when the Broker is run in a deployment with a long `terminationGracePeriodSeconds` pods take an extremely long time to tear down.

This is because the K8s lifecycle first sends a SIGTERM, which is a warning shot to tell the application to tear down gracefully, after which it is given `terminationGracePeriodSeconds` to exit before it is sent `SIGKILL`.  I ran this in a pod with a `terminationGracePeriodSeconds` of several minutes and pods took forever to tear down, but with this fix it takes roughly the 60s of the sleep below this logic.