Remove Implementation Specific Checks From Tracing Conformance Tests (#2837)
* Directly shadow t in tracing helpers

* Remove artificial root spans from test trees.

* Support brokerclass flag in the tracing conformance tests

* Remove Root field and unused SpanCount method from TestSpanTree

* Reserve variable name t for testing.T in TestSpanTree methods

* Do not capitalize error messages returned by TestSpanTree methods

* Show error when a Span fails to match in TestSpanTree.MatchesSubtree

* Assert only broker spans at the ingress and filter boundaries

This prevents the conformance test from asserting internal details of
the broker implementation and allows brokers to emit extra spans as
necessary.

* Refactor tracing tests to not check broker ingress address

A fixed broker ingress address cannot be used since the address is
implementation dependent. It would be possible to retrieve the address
from the broker status after the broker is healthy but it is not clear
that this is an important aspect of broker tracing conformance.

* Enable tracing conformance tests for MTChannelBasedBroker

* Assert local endpoint service names in the broker tracing test

* Use W3C trace propagation format in mtbroker ingress and filter

* Do not create config map propagation when testing mtbroker