Remove strict subscriber feature flag (#6608)
Fixes #5756

As per https://github.com/knative/specs/pull/30 the spec was changed
from requiring one of `spec.subscriber` or `spec.reply` to be set to
require that `spec.subscriber` is set.

As part of this change the `strict-subscriber` feature flag was
introduced to disable this validation (#5762) and promoted to beta
(enabled by default) in 1.7 (#6473).

:broom: This commit removes this feature flag again so the validation
can't be disabled anymore.

## Proposed Changes
* Remove `strict-subscriber` feature flag
* Promote strict subscriber validation to GA

### Pre-review Checklist

- [ ] **At least 80% unit test coverage**
- [ ] **E2E tests** for any new behavior
- [ ] **Docs PR** for any user-facing impact
- [ ] **Spec PR** for any new API feature
- [ ] **Conformance test** for any change to the spec

**Release Note**
```release-note
Remove the possibility to disable strict subscriber validation.
When the reply field is specified without a subscriber, the reply field won't be used as a subscriber by default and the subscriber validation will fail.
```

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>