Use `AddressStatus` instead of `Addressable` directly in broker status (#6622)
Currently the BrokerStatus field uses the `address` directly:

https://github.com/knative/eventing/blob/41653ce64478d0a74470d4609510f2dc3c611de6/pkg/apis/eventing/v1/broker_types.go#L84-L99

Anyhow there is an `AddressStatus` struct, which encapsulates this and
(according to its docs) should be used to embed an addressable:

https://github.com/knative/eventing/blob/41653ce64478d0a74470d4609510f2dc3c611de6/vendor/knative.dev/pkg/apis/duck/v1/addressable_types.go#L64-L66

E.g. Channels use this struct:

https://github.com/knative/eventing/blob/41653ce64478d0a74470d4609510f2dc3c611de6/pkg/apis/duck/v1/channelable_types.go#L56-L75

## Proposed Changes
- :broom: Change the BrockerStatus struct to use the `AddressStatus`
struct.

:exclamation: As this struct will be embedded inline, the JSON/YAML
structure should not change. Anyhow the Golang API will change, as the
`Address` field from the `AddressStatus` can be `nil` (before this
change it couldn't).
I am not sure, if this will break any users of `pkg/apis/eventing` and
we shouldn't change this in `v1`, so feedback is highly welcome.

### Pre-review Checklist
- [ ] **At least 80% unit test coverage**
- [ ] **E2E tests** for any new behavior
- [ ] **Docs PR** for any user-facing impact
- [ ] **Spec PR** for any new API feature
- [ ] **Conformance test** for any change to the spec

**Release Note**
```release-note
```

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>