Updated DEVELOPMENT.md to provide better instructions on setting up kubernetes (#6977)
# Fixes

<!-- Please include the 'why' behind your changes if no issue exists -->
When following `DEVELOPMENT.md` to get set up to develop on this
project, I noticed that the section on setting up a kubernetes cluster
linked to a page which didn't seem to cover how to set up a kubernetes
cluster.

## Proposed Changes

<!-- Please categorize your changes:
- :gift: Add new feature
- :bug: Fix bug
- :broom: Update or clean up current behavior
- :wastebasket: Remove feature or internal logic
-->

- Update `DEVELOPMENT.md` to provide more helpful links to setting up a
kubernetes cluster locally


### Pre-review Checklist

<!-- If these boxes are not checked, you will be asked to complete these
requirements or explain why they do not apply to your PR. -->

- [ ] **At least 80% unit test coverage**
- [ ] **E2E tests** for any new behavior
- [ ] **Docs PR** for any user-facing impact
- [ ] **Spec PR** for any new API feature
- [ ] **Conformance test** for any change to the spec

As this PR only updates the development setup docs, I don't think these
requirements apply to it.

**Release Note**

<!--
:page_facing_up: If this change has user-visible impact, write a release
note in the block
below. Include the string "action required" if additional action is
required of
users switching to the new release, for example in case of a breaking
change.

Write as if you are speaking to users, not other Knative contributors.
If this
change has no user-visible impact, no release note is needed.
-->

```release-note

```


**Docs**

<!--
:book: If this change has user-visible impact, link to an issue or PR in
https://github.com/knative/docs.
-->